requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
PyMuPDF
litellm
trulens
trulens-apps-llamaindex
torch 
torchvision 
torchaudio 
